[Fixes] Fixed the prototype for reversed function, it should return an unsigned long int

+++ 0x08-palindrome_integer/palindrome.h
@@ -1,7 +1,7 @@
 #ifndef PALINDROME_H
 #define PALINDROME_H
 
-int is_palindrome(unsigned long n);
+unsigned long is_palindrome(unsigned long n);
 int reversed(unsigned long n);
 
 #endif
